Moved monitor of market hours in strategy - cleanup imports
