+++ scripts/StocksAutoTrader.py
@@ -1,10 +1,6 @@
 import logging
-import sys
-import pytz
-import datetime
-import time
+import json
 
-from Utils import *
 from IGInterface import IGInterface
 from Strategies.SimpleMACD import SimpleMACD
 
@@ -17,7 +13,6 @@ class StocksAutoTrader:
         self.strategy = SimpleMACD(config)
 
     def read_configuration(self, config):
-        self.time_zone = config['general']['time_zone']
         self.epic_ids_filepath = config['general']['epic_ids_filepath']
 
     def get_epic_ids(self):
@@ -30,7 +25,6 @@ class StocksAutoTrader:
                 for line in filecontents:
                     # remove linebreak which is the last character of the string
                     current_epic_id = line[:-1]
-                    # add item to the list
                     epic_ids.append(current_epic_id)
         except IOError:
             # Create the file empty
@@ -38,11 +32,6 @@ class StocksAutoTrader:
             filehandle = open(self.epic_ids_filepath, 'w')
         logging.debug(epic_ids)
         return epic_ids
-
-    def isMarketOpen(self, timezone):
-        tz = pytz.timezone(timezone)
-        now_time = datetime.datetime.now(tz=tz).strftime('%H:%M')
-        return is_between(str(now_time), ("07:55", "16:35"))
 
     def start(self, argv):
         # Read credentials file
@@ -61,11 +50,4 @@ class StocksAutoTrader:
         # Read list of company epic ids
         main_epic_ids = self.get_epic_ids()
 
-        while True:
-            if not self.isMarketOpen(self.time_zone):
-                logging.info("Market is closed! Wait...")
-                time.sleep(60)
-                continue
-            else:
-                self.strategy.spin(self.IG, main_epic_ids)
-                time.sleep(1)
+        self.strategy.start(self.IG, main_epic_ids)
